requirements.txt
discord.py==2.6.2
python-dotenv==1.1.1
aiohttp==3.12.15
Flask==3.1.0

